chore(engine): add TestWatcher to other PropertyBasedTest
